moved all auxiliary classes from cals_dose_logic.py to separate files
