+++ cals_dose_logic.py
@@ -7,686 +7,33 @@ import json
 import ctypes
 import matplotlib.pyplot as plt
 import numpy as np
-import pandas
-import tifffile as tifimage
-import matplotlib.widgets
-from PyQt5.QtCore import pyqtSignal, QThread
+from PyQt5.QtCore import pyqtSignal
 from PyQt5.QtGui import QIcon
-from scipy.optimize import curve_fit
-from PyQt5 import QtWidgets, QtCore
+from PyQt5 import QtWidgets
 from PyQt5.QtWidgets import QFileDialog, QLineEdit, QDoubleSpinBox, QMessageBox, QCheckBox
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
 from matplotlib.widgets import RectangleSelector
 from Dose import Ui_MainWindow
-from calibrate_list import Ui_Form
-from Axes import Ui_Form as Axes_form
-from Curve import Ui_Form as Curve_form
-from Values import Ui_Form as Values_form
-from DB_and_settings import Ui_Form as DB_form
 from database import db_connection
-from database import dbProxy as db
-from doses_and_pathes import DosesAndPaths
-from logicParser import LogicODVariant, LogicCurveVariants, LogicCurveFitsVariant, LogicParser
+from logicParser import LogicParser
 from filters import Filters, Filter
-from stats import stats_ui
+from DosesAndPaths import DosesAndPaths
+from GraphicsPlotting import GraphicsPlotting
+import GraphicsPlotting as gp
+from DoseClass import Dose
+from SaveLoadData import SaveLoadData
+import SaveLoadData as sld
+from DatabaseAndSettings import DatabaseAndSettings
+import DatabaseAndSettings as das
+from Warnings import Warnings
+from ValuesWindow import ValuesWindow
+from CurveWindow import CurveWindow
+from Form import Form
+import Form as fm
+from AxesWindow import AxesWindow
 
 plt.switch_backend('agg')
-
-
-class GraphicsPlotting:
-    @staticmethod
-    def draw_dose_map(z):
-        """
-        This method draws a map of dose distribution
-        """
-        application.figure_map.clf()
-        ax = application.figure_map.add_subplot(111)
-        application.cursor = matplotlib.widgets.Cursor(ax, useblit=False, color='red', linewidth=1)
-        application.canvas_map.mpl_connect('button_press_event', lambda event: application.onclick(event, ax))
-        application.canvas_map.draw()
-
-        if DosesAndPaths.vmax is not None and DosesAndPaths.vmin is not None:
-            im3 = ax.imshow(z, cmap="jet", vmin=DosesAndPaths.vmin, vmax=DosesAndPaths.vmax)
-        else:
-            im3 = ax.imshow(z, cmap="jet")
-
-        formatter = lambda x, pos: round(x * DosesAndPaths.basis_formatter, 0)  # the resolution
-        ax.xaxis.set_major_formatter(formatter)
-        ax.yaxis.set_major_formatter(formatter)
-        application.figure_map.colorbar(im3, ax=ax, orientation="vertical")
-        application.canvas_map.draw()
-
-    @staticmethod
-    def draw_curve(func, calculation_doses, setting_doses, p_opt, figure_graph, canvas_graph, sigma):
-        """
-        This method draws dose curve
-        """
-        figure_graph.clf()
-        ax = figure_graph.add_subplot(111)
-        ax.errorbar(calculation_doses, setting_doses, yerr=np.array(setting_doses[0:]) * (sigma / 100), fmt='ro',
-                    label="Data points", markersize=6, capsize=5)
-        ax.plot(calculation_doses, func(calculation_doses, *p_opt), label="Fit function", color="black", linestyle="-.")
-        ax.grid(True, linestyle="-.")
-        ax.legend(loc="best")
-        ax.set_ylabel('Absorbed dose, Gy')
-        ax.set_xlabel('Relative optical density')
-        ax.set_xlim(0, np.max(calculation_doses) + 0.015)
-        ax.set_ylim(0, np.max(setting_doses) + 0.5)
-        canvas_graph.draw()
-
-    @staticmethod
-    def draw_curve_from_db(doses, ods, dose_object, figure_graph, canvas_graph):
-        figure_graph.clf()
-        ax = figure_graph.add_subplot(111)
-        ax.plot(ods, doses, '*', label="Data points", color="red")
-        ax.plot(np.linspace(ods[0], ods[-1], 500), dose_object.evaluateOD(np.linspace(ods[0], ods[-1], 500)),
-                label="Fit function", color="black")
-        ax.grid(True, linestyle="-.")
-        ax.legend(loc="best")
-        ax.set_ylabel('Absorbed dose, Gy')
-        ax.set_xlabel('Relative optical density')
-        canvas_graph.draw()
-
-
-class Dose(QThread):
-    """
-    Calculate dose
-    """
-    progressChanged = QtCore.pyqtSignal(int)
-
-    def __init__(self, zero_dose, zero_dose_for_irrad_film, calibrate_list, doses_list, irradiation_film, sigma,
-                 func_name):
-        super().__init__()
-        self.zero_dose = zero_dose
-        self.zero_dose_for_irrad_film = zero_dose_for_irrad_film
-        self.calibrate_list = calibrate_list
-        self.irradiation_film = irradiation_film
-        self.setting_doses = doses_list
-        self.sigma = sigma
-        self.func_name = func_name
-
-    def run(self):
-        """
-        Start thread
-        """
-        self.red_channel_calc()
-        self.calculate_calibrate_film()
-        self.calc_dose_map()
-
-    @staticmethod
-    def fit_func(func_name):
-        """
-        Returns the static method defining the approximation function
-        :param func_name: method name
-        """
-        functions = {
-            'base fit_func': Dose.fit_func1,
-            'fit_func_pol2': Dose.fit_func_pol2,
-            'fit_func_pol3': Dose.fit_func_pol3,
-            'fit_func_pol5': Dose.fit_func_pol5
-        }
-        if func_name in functions:
-            return functions[func_name]
-
-    @staticmethod
-    def fit_func1(od, a, b, c):
-        """
-        Fitting function for calibration curve
-        :param od:
-        :param a:
-        :param b:
-        :param c:
-        :return:
-        """
-        return (b / (od - a)) + c
-
-    @staticmethod
-    def fit_func_pol2(od, x2, x1, x0):
-        '''
-        Fit with x2*od**2+x1*od+x0
-        '''
-        func = np.poly1d([x2, x1, x0])
-        return func(od)
-
-    @staticmethod
-    def fit_func_pol3(od, x3, x2, x1, x0):
-        '''
-        Fit with x3*od**3+x2*od**2+x1*od+x0
-        '''
-        func = np.poly1d([x3, x2, x1, x0])
-        return func(od)
-
-    @staticmethod
-    def fit_func_pol5(od, x5, x4, x3, x2, x1, x0):
-        '''
-        Fit with x5*od**5+x4*od**4+x3*od**3+x2*od**2+x1*od+x0
-        '''
-        func = np.poly1d([x5, x4, x3, x2, x1, x0])
-        return func(od)
-
-    @staticmethod
-    def get_imarray(img):
-        im = tifimage.imread(img)
-        imarray = np.array(im, dtype=np.uint16)
-        imarray = (imarray[:, :, 0])
-        return imarray
-
-    def red_channel_calc(self):
-        """
-        Calculate value in red channel of blank field
-        :return: od_blank
-        """
-        blank_field_path = self.zero_dose
-        od_blank = np.mean(Dose.get_imarray(blank_field_path))
-        print("Blank field value: ", round(od_blank, 2))
-
-        DosesAndPaths.red_channel_blank = od_blank
-        return od_blank
-
-    def calc_dose(self, path_to_film):
-        """
-        Image file processor
-        """
-        red_channel_current_tiff = np.mean(Dose.get_imarray(path_to_film))
-        od_current_dose = np.log10(DosesAndPaths.red_channel_blank / red_channel_current_tiff)
-
-        return od_current_dose
-
-    def find_best_fit(self, path_to_film):
-        """
-        Finding best fit dose and writes it to the list
-        """
-        od_current_dose = self.calc_dose(path_to_film) - self.zero_dose
-        DosesAndPaths.calculation_doses.append(od_current_dose)
-
-    def calculate_calibrate_film(self):
-        """
-        Calculating dose for each file
-        """
-        # сначала считаем нулевую дозу
-        self.zero_dose = self.calc_dose(list(self.calibrate_list)[0])
-        # затем считаем для каждого файла с использованием посчитанной нулевой
-        for i in self.calibrate_list:
-            self.find_best_fit(i)
-        try:
-            p_opt, p_cov = curve_fit(self.fit_func(self.func_name), np.array(DosesAndPaths.calculation_doses[1:]),
-                                     np.array(self.setting_doses[1:]),
-                                     sigma=np.array(self.setting_doses[1:]) * (self.sigma / 100))
-            DosesAndPaths.p_opt = p_opt
-        except (ValueError, RuntimeError):
-
-            print('Incorrect sigma value')
-
-    def calc_dose_map(self):
-        """
-        Working with user image
-        """
-        try:
-            zero_dose_for_irrad_film = self.calc_dose(self.zero_dose_for_irrad_film)
-            print("\nShape of scanned film:", np.shape(CalcUI.choose_orig_or_crop()))
-            progress = 0
-            counter = 0
-            print("\nPrepearing your file:\n")
-            for i in np.nditer(CalcUI.choose_orig_or_crop()):
-                x = np.log10(DosesAndPaths.red_channel_blank / i)
-                x = x - zero_dose_for_irrad_film
-                x = self.fit_func(self.func_name)(x, *DosesAndPaths.p_opt)
-                DosesAndPaths.z = np.append(DosesAndPaths.z, x)
-
-                counter = counter + 1
-                if counter % 10000 == 0:
-                    print("Iteration ", counter, "/", np.size(CalcUI.choose_orig_or_crop()))
-                    progress += 1
-                    self.progressChanged.emit(round(progress))
-
-            DosesAndPaths.z = DosesAndPaths.z.reshape(np.shape(CalcUI.choose_orig_or_crop()))
-            print("\nDose calculation ended!!!\n")
-            self.progressChanged.emit(100)
-            GraphicsPlotting().draw_dose_map(DosesAndPaths.z)
-        except ValueError:
-            print('No files found')
-
-
-class Form(QtWidgets.QWidget, Ui_Form):
-    """
-    UI class for displaying dose/path list
-    """
-
-    def __init__(self, *args, **kwargs):
-        QtWidgets.QWidget.__init__(self, *args, **kwargs)
-        self.setupUi(self)
-        self.widget_count = 0
-        self.all_widgets = None
-        self.curve_win = None
-        self.value_win = None
-
-        self.pushButton_5.setDisabled(True)
-        self.pushButton_9.setDisabled(True)
-
-        self.pushButton.clicked.connect(lambda: self.get_empty_field_file(self.lineEdit))
-        self.pushButton_2.clicked.connect(self.dynamic_add_fields)
-        self.pushButton_3.clicked.connect(self.dynamic_delete_fields)
-        self.pushButton_4.clicked.connect(self.get_all_params_widgets)
-        self.pushButton_5.clicked.connect(self.draw_curve)
-        self.pushButton_8.clicked.connect(SaveLoadData.create_json)
-        self.pushButton_6.clicked.connect(lambda: self.get_empty_field_file(self.lineEdit_2))
-        self.pushButton_9.clicked.connect(self.get_values)
-
-    def search_file(self):
-        """
-        Search file
-        """
-        file_name = \
-            QFileDialog.getOpenFileName(self, 'Open file', '', '*.tif', None, QFileDialog.DontUseNativeDialog)[0]
-        return file_name
-
-    def get_empty_field_file(self, line):
-        """Set the found file in lineEdit"""
-        line.setText(self.search_file())
-
-        if len(line.text()) != 0:
-            line.setDisabled(True)
-
-    def dynamic_add_fields(self):
-        """
-        Dynamically adding fields to record new paths/doses
-        """
-        spin_box = QtWidgets.QDoubleSpinBox()
-        qline_edit = QtWidgets.QLineEdit()
-        push_button = QtWidgets.QPushButton("Select")
-        checkbox = QtWidgets.QCheckBox()
-
-        self.gridLayout_3.addWidget(checkbox)
-        self.gridLayout_3.addWidget(spin_box)
-        self.gridLayout_3.addWidget(qline_edit)
-        self.gridLayout_3.addWidget(push_button)
-
-        push_button.clicked.connect(lambda: self.get_empty_field_file(qline_edit))
-
-        self.widget_count += 1
-        self.adjustSize()
-
-    def dynamic_delete_fields(self):
-        """
-        Deleting fields to record paths/doses
-        """
-        widgets = [self.gridLayout_3.itemAt(i).widget() for i in range(self.gridLayout_3.count())]
-        checkboxes = [i for i in widgets if isinstance(i, QCheckBox)]
-
-        for k in checkboxes:
-            if k.checkState() == 2:
-                index = widgets.index(k)
-                for j in range(4):
-                    my_widget = self.gridLayout_3.itemAt(index).widget()
-                    my_widget.setParent(None)
-                    self.adjustSize()
-                    widgets.remove(my_widget)
-
-    def get_all_params_widgets(self):
-        """
-        Set values from fields to variables
-        """
-        doses = []
-        paths = []
-        sigma = self.spinBox.value()
-        widgets = (self.gridLayout_3.itemAt(i).widget() for i in range(self.gridLayout_3.count()))
-        self.all_widgets = widgets
-        for widget in widgets:
-            if isinstance(widget, QLineEdit):
-                paths.append(widget.text())
-            if isinstance(widget, QDoubleSpinBox):
-                doses.append(widget.value())
-            if isinstance(widget, QLineEdit) and widget.text() is '':
-                Warnings.error_empty_value()
-
-        DosesAndPaths.doses = doses
-        DosesAndPaths.paths = paths
-        DosesAndPaths.sigma = sigma
-        DosesAndPaths.fit_func_type = self.comboBox.currentText()
-        DosesAndPaths.empty_scanner_field_file = self.lineEdit_2.text()
-        DosesAndPaths.calculation_doses.clear()
-
-        self.get_enabled_curve_drawing()
-        CalcUI.HAND_SWITCH_MODE = True
-
-    def get_enabled_curve_drawing(self):
-        """
-        Blocking the availability of fields for editing
-        """
-        if len(DosesAndPaths.doses) is not 0 and len(DosesAndPaths.paths) is not 0:
-            self.pushButton_5.setDisabled(False)
-
-    def draw_curve(self):
-        """
-        The method connects methods of drawing graphs
-        """
-        self.curve_win = CurveWindow()
-        self.curve_win.get_curve()
-        self.curve_win.setMinimumSize(640, 480)
-        self.curve_win.show()
-        self.pushButton_5.setDisabled(True)
-        self.pushButton_9.setDisabled(False)
-
-    def create_widgets_second_open(self):
-        """
-        The method restores the fields created by the user when the window is reopened
-        """
-        data_count = len(DosesAndPaths.doses)
-        if self.gridLayout_3.count() >= 6 and data_count > 1:
-            for i in range(data_count - 1):
-                self.dynamic_add_fields()
-
-    def insert_data_in_fields(self):
-        """
-        Fill in the dialog window with doses and paths when opening
-        """
-        widgets = [self.gridLayout_3.itemAt(i).widget() for i in range(self.gridLayout_3.count())]
-        lineedits = [i for i in widgets if isinstance(i, QLineEdit)]
-        spinboxes = [i for i in widgets if isinstance(i, QDoubleSpinBox)]
-
-        for path, dose, line, spin in zip(DosesAndPaths.paths, DosesAndPaths.doses, lineedits, spinboxes):
-            if isinstance(line, QLineEdit):
-                line.setText(path)
-            if isinstance(spin, QDoubleSpinBox):
-                spin.setValue((float(dose)))
-
-        self.spinBox.setValue(DosesAndPaths.sigma)
-        self.lineEdit_2.setText(DosesAndPaths.empty_scanner_field_file)
-        self.comboBox.setCurrentText(DosesAndPaths.fit_func_type)
-
-    def get_values(self):
-        """
-        Get the values of doses, calculation doses and p_opt in dialog window
-        """
-        self.value_win = ValuesWindow()
-        if len(DosesAndPaths.calculation_doses) > 0:
-            self.value_win.plainTextEdit.appendPlainText(
-                ('DOSES: \n' + ('\n'.join(map(str, [round(x, 4) for x in DosesAndPaths.doses]))).replace('.', ',')))
-            self.value_win.plainTextEdit.appendPlainText(('\nOPTICAL DENSITY: \n' + (
-                '\n'.join(map(str, [round(x, 4) for x in DosesAndPaths.calculation_doses]))).replace('.', ',')))
-            self.value_win.plainTextEdit.appendPlainText(('\nPOLY_COEF_A_B_C: \n' + (
-                '\n'.join(map(str, [round(x, 4) for x in DosesAndPaths.p_opt]))).replace('.', ',')))
-        self.value_win.show()
-
-
-class ValuesWindow(QtWidgets.QWidget, Values_form):
-    """
-    Class of the dialog window with a values
-    """
-
-    def __init__(self, *args, **kwargs):
-        QtWidgets.QWidget.__init__(self, *args, **kwargs)
-        self.setupUi(self)
-
-
-class CurveWindow(QtWidgets.QWidget, Curve_form):
-    """
-    Class of the dialog window with a curve
-    """
-    closeDialog = pyqtSignal()
-
-    def __init__(self, *args, **kwargs):
-        QtWidgets.QWidget.__init__(self, *args, **kwargs)
-        self.setupUi(self)
-
-        self.figure_graph = plt.figure()
-        self.canvas_graph = FigureCanvas(self.figure_graph)
-        self.toolbar_graph = NavigationToolbar(self.canvas_graph, self)
-        self.verticalLayout_5.addWidget(self.toolbar_graph)
-        self.verticalLayout_7.addWidget(self.canvas_graph)
-
-    def get_curve(self):
-        """
-        Draw dose curve
-        """
-        try:
-            calc = Dose(DosesAndPaths.empty_scanner_field_file, DosesAndPaths.empty_field_file, DosesAndPaths.paths,
-                        DosesAndPaths.doses,
-                        DosesAndPaths.irrad_film_file, DosesAndPaths.sigma, DosesAndPaths.fit_func_type)
-            calc.red_channel_calc()
-            calc.calculate_calibrate_film()
-            GraphicsPlotting.draw_curve(Dose.fit_func(DosesAndPaths.fit_func_type), DosesAndPaths.calculation_doses,
-                                        DosesAndPaths.doses, DosesAndPaths.p_opt, self.figure_graph, self.canvas_graph,
-                                        DosesAndPaths.sigma)
-        except (ValueError, TypeError):
-            print('Incorrect parameters')
-
-    def get_curve_from_db_data(self, doses, ods, evaluate_od):
-        """
-        Draw dose curve from db data
-        """
-        try:
-            GraphicsPlotting.draw_curve_from_db(doses[1:], ods[1:], evaluate_od, self.figure_graph, self.canvas_graph)
-        except Exception as e:
-            Warnings.error_incorrect_value()
-            template = "An exception of type {0} occurred. Arguments:\n{1!r}"
-            message = template.format(type(e).__name__, e.args)
-            print(message)
-
-    def closeEvent(self, event):
-        """
-        Clear figure
-        :param event: Window close
-        """
-        plt.close(self.figure_graph)
-        self.closeDialog.emit()
-
-
-class MoveGraphLine(QtWidgets.QWidget):
-    """
-    Adds functionality to move graphs along the X-axis.
-    """
-    dataChanged = pyqtSignal(np.ndarray, np.ndarray)
-
-    def __init__(self, graf, ax, move_speed=0.5):
-        QtWidgets.QWidget.__init__(self)
-        self.graf = graf  # The graph to be moved
-        self.ax = ax  # The axes object of the graph
-        self.figcanvas = self.ax.figure.canvas  # The canvas of the figure
-        self.start_point = None  # The starting point of the mouse drag
-        self.move_speed = move_speed  # The speed of the graph movement
-        self.moving = False  # Flag indicating whether the graph is moving or not
-
-        # Connect the mouse events to their respective handlers
-        self.figcanvas.mpl_connect('button_press_event', self.mouse_press)
-        self.figcanvas.mpl_connect('button_release_event', self.mouse_release)
-        self.figcanvas.mpl_connect('motion_notify_event', self.mouse_move)
-
-    def mouse_release(self, event):
-        # Check if a navigation tool is active, if so, do nothing
-        if self.ax.get_navigate_mode() is not None:
-            return
-        self.moving = False
-
-    def mouse_press(self, event):
-        # Check if a navigation tool is active, if so, do nothing
-        if self.ax.get_navigate_mode() is not None:
-            return
-        if event.inaxes != self.ax:
-            return
-        self.start_point = (event.xdata, event.ydata)
-        self.moving = True
-
-    def mouse_move(self, event):
-        # Check if a navigation tool is active, if so, do nothing
-        if self.ax.get_navigate_mode() is not None:
-            return
-        if event.inaxes != self.ax:
-            return
-        if not self.moving:
-            return
-
-        # Calculate the shift in the x-axis based on the mouse movement
-        shift_x = (self.start_point[0] - event.xdata) * self.move_speed
-        mvdx, mvdy = self.graf.get_data()  # Get the current graph data
-        mvdx -= shift_x  # Update the x-axis data by applying the shift
-
-        # Update the graph with the new data and redraw the canvas
-        self.graf.set_data(mvdx, mvdy)
-        self.figcanvas.draw()
-
-        # Emit the dataChanged signal with the updated data
-        self.dataChanged.emit(mvdx, mvdy)
-
-
-class AxesWindow(QtWidgets.QWidget, Axes_form):
-    """
-    Class for drawing graphs on the X and Y axes
-    """
-    closeDialog = pyqtSignal()
-    dataChangedLeft = pyqtSignal(str)
-    dataChangedRight = pyqtSignal(str)
-
-    def __init__(self, *args, **kwargs):
-        QtWidgets.QWidget.__init__(self, *args, **kwargs)
-        self.setupUi(self)
-        self.value_win = None
-
-        self.figure_map_x = plt.figure()
-        self.canvas_map_x = FigureCanvas(self.figure_map_x)
-        self.verticalLayout.addWidget(self.canvas_map_x)
-        self.toolbar_x = NavigationToolbar(self.canvas_map_x, self)
-        self.verticalLayout_4.addWidget(self.toolbar_x)
-
-        self.figure_map_y = plt.figure()
-        self.canvas_map_y = FigureCanvas(self.figure_map_y)
-        self.verticalLayout_3.addWidget(self.canvas_map_y)
-        self.toolbar_y = NavigationToolbar(self.canvas_map_y, self)
-        self.verticalLayout_5.addWidget(self.toolbar_y)
-
-        self.formatted_mvdx_x = None
-        self.formatted_mvdx_y = None
-
-        # Stats windows
-        self.stats_left = stats_ui.MainWindow(self.pushButton_5, self, position='left')
-        self.stats_right = stats_ui.MainWindow(self.pushButton_6, self, position='right')
-
-        # Stats buttons
-        self.pushButton_5.clicked.connect(self.on_button_left_clicked)
-        self.pushButton_6.clicked.connect(self.on_button_right_clicked)
-
-        self.ax_x = None
-        self.ax_y = None
-
-    @staticmethod
-    def dose_limits_for_graph(slice, ax):
-        """
-        Limits the dose if a checkbox was checked
-        :param slice: axis segment
-        :param ax: сanvas
-        :return: slice
-        """
-        if DosesAndPaths.vmax is not None and DosesAndPaths.vmin is not None:
-            slice_clipped = np.clip(slice, DosesAndPaths.vmin, DosesAndPaths.vmax)
-            graf, = ax.plot(slice_clipped)
-            return graf, slice_clipped
-        else:
-            graf, = ax.plot(slice)
-            return graf, slice
-
-    def draw_graphics(self, slice_x, slice_y):
-        """
-        Method for graphing the X and Y axes
-        :param slice_x: Set of values along the X-axis
-        :param slice_y: Set of values along the Y-axis
-        """
-        formatter = lambda x, pos: round(x * DosesAndPaths.basis_formatter, 0)  # the resolution
-
-        # x axis
-        self.figure_map_x.clf()
-        self.ax_x = self.figure_map_x.add_subplot(111)
-        ax_x = self.ax_x
-        ax_x.grid(True, linestyle="-.")
-        graf_x, slice_values_x = AxesWindow.dose_limits_for_graph(slice_x, ax_x)
-        ax_x.xaxis.set_major_formatter(formatter)
-        ax_x.set_xlabel('mm')
-        ax_x.set_ylabel('Absorbed dose, Gy')
-        self.canvas_map_x.draw()
-        self.pushButton.clicked.connect(lambda: self.get_values(slice_values_x, 'X axis'))
-        self.pushButton_2.clicked.connect(lambda: SaveLoadData.save_as_excel_file_axis(slice_values_x,
-                                                                                       'X axis', self.formatted_mvdx_x))
-
-        # Add a function to move the X diagram along the X axis
-        self.moveline_x_x = MoveGraphLine(graf_x, ax_x, move_speed=0.1)
-        self.moveline_x_x.dataChanged.connect(self.handle_data_changed_x_x)
-
-        # y axis
-        self.figure_map_y.clf()
-        self.ax_y = self.figure_map_y.add_subplot(111)
-        ax_y = self.ax_y
-        ax_y.grid(True, linestyle="-.")
-        graf_y, slice_values_y = AxesWindow.dose_limits_for_graph(slice_y, ax_y)
-        ax_y.xaxis.set_major_formatter(formatter)
-        ax_y.set_xlabel('mm')
-        ax_y.set_ylabel('Absorbed dose, Gy')
-        self.canvas_map_y.draw()
-        self.pushButton_3.clicked.connect(lambda: self.get_values(slice_values_y, 'Y axis'))
-        self.pushButton_4.clicked.connect(lambda: SaveLoadData.save_as_excel_file_axis(slice_values_y,
-                                                                                       'Y axis', self.formatted_mvdx_y))
-        # Add a function to move the Y diagram along the X axis
-        self.moveline_x_y = MoveGraphLine(graf_y, ax_y, move_speed=0.1)
-        self.moveline_x_y.dataChanged.connect(self.handle_data_changed_x_y)
-
-        # Save ours data in dataclass
-        self.get_final_params_for_stats(slice_values_x, slice_values_y)
-
-    def get_final_params_for_stats(self, slice_values_x, slice_values_y):
-        # Function for save parameters in dataclass
-        DosesAndPaths.final_slice_values_x = slice_values_x
-        DosesAndPaths.final_slice_values_y = slice_values_y
-
-    def get_values(self, values, ax_name):
-        """
-        Get the values of doses on the axis in dialog window
-        :param values: values of doses
-        :param ax_name: name of the axis
-        """
-        self.value_win = ValuesWindow()
-        self.value_win.label.setText(ax_name)
-        values_with_separator = ('\n'.join(map(str, [round(x, 4) for x in values]))).replace('.', ',')
-        if len(values) > 0:
-            self.value_win.plainTextEdit.appendPlainText(values_with_separator)
-        self.value_win.show()
-
-    def closeEvent(self, event):
-        """
-        Clear figure
-        :param event: Window close
-        """
-        plt.close(self.figure_map_x)
-        plt.close(self.figure_map_y)
-        self.closeDialog.emit()
-
-    def handle_data_changed_x_x(self, mvdx, mvdy):
-        # Multiply the raw x-axis data (mvdx) by the basis
-        # formatter to convert the values to the desired scale
-        formatted_mvdx = mvdx * DosesAndPaths.basis_formatter
-
-        # Round the formatted x-axis data to 0 decimal
-        # places and store it as an attribute of the class
-        self.formatted_mvdx_x = formatted_mvdx.round(2)
-        DosesAndPaths.final_formatted_mvdx_x = self.formatted_mvdx_x
-        self.dataChangedLeft.emit('left')
-
-    def handle_data_changed_x_y(self, mvdx, mvdy):
-        # Multiply the raw x-axis data (mvdx) by the basis
-        # formatter to convert the values to the desired scale
-        formatted_mvdx = mvdx * DosesAndPaths.basis_formatter
-
-        # Round the formatted x-axis data to 0 decimal
-        # places and store it as an attribute of the class
-        self.formatted_mvdx_y = formatted_mvdx.round(2)
-        DosesAndPaths.final_formatted_mvdx_y = self.formatted_mvdx_y
-        self.dataChangedRight.emit('right')
-
-    def on_button_left_clicked(self):
-        # Show panel with stats on X
-        self.stats_left.show_panel()
-
-    def on_button_right_clicked(self):
-        # Show panel with stats on Y
-        self.stats_right.show_panel()
 
 
 class CalcUI(QtWidgets.QMainWindow):
@@ -951,13 +298,13 @@ class CalcUI(QtWidgets.QMainWindow):
         Defining the first file from the calibration as unexposed
         :return:
         """
-        if self.ui.checkBox.isChecked() and CalcUI.HAND_SWITCH_MODE:
+        if self.ui.checkBox.isChecked() and application.HAND_SWITCH_MODE:
             try:
                 DosesAndPaths.empty_field_file = DosesAndPaths.paths[0]
             except (TypeError, IndexError):
                 Warnings.error_confirm_calibration()
 
-        elif self.ui.checkBox.isChecked() and not CalcUI.HAND_SWITCH_MODE:
+        elif self.ui.checkBox.isChecked() and not application.HAND_SWITCH_MODE:
             DosesAndPaths.empty_field_file = DosesAndPaths.zero_from_db
         elif not self.ui.checkBox.isChecked() and len(self.ui.lineEdit.text()) is not 0:
             DosesAndPaths.empty_field_file = self.ui.lineEdit.text()
@@ -980,11 +327,11 @@ class CalcUI(QtWidgets.QMainWindow):
 
         self.first_film_from_calibration()
 
-        if self.check_fields_manual_mode() and CalcUI.HAND_SWITCH_MODE:
+        if self.check_fields_manual_mode() and application.HAND_SWITCH_MODE:
             # manual mode
             self.calc_from_manual()
 
-        if self.check_fields_bd_mode() and not CalcUI.HAND_SWITCH_MODE:
+        if self.check_fields_bd_mode() and not application.HAND_SWITCH_MODE:
             # db mode
             if self.ui.checkBox.isChecked():
                 empty_file = DosesAndPaths.zero_from_db
@@ -1094,27 +441,8 @@ class CalcUI(QtWidgets.QMainWindow):
             return True
 
 
-class Warnings:
-    """
-    This class contains pop-up errors
-    """
 
-    @staticmethod
-    def error_exist_files(file: list):
-        """
-        :param file: list only
-        :return:
-        """
-        QMessageBox.critical(None, "Error", "<b>No such file or directory</b><br><br>"
-                                            f"Some selected files are not on the disk. List of files:<br>"
-                                            f"{'<br>'.join(file)}", QMessageBox.Ok)
 
-    @staticmethod
-    def error_special_symbols():
-        QMessageBox.critical(None, "Error ", "<b>Incorrect name</b><br><br>"
-                                             "Please re-save the file using the correct name without special "
-                                             "characters",
-                             QMessageBox.Ok)
 
     @staticmethod
     def error_incorrect_value():
@@ -1612,6 +940,10 @@ if __name__ == "__main__":
     app.setWindowIcon(app_icon)
     application = CalcUI()
     # win title
+    gp.application = application
+    sld.application = application
+    das.application = application
+    fm.application = application
     application.setWindowTitle("Dose calculator")
     # set minimum size
     application.setMinimumSize(1200, 800)
